CodeExecRAGv1/requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
exceptiongroup==1.2.1
FLAML==2.1.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
lxml==5.2.2
numpy==1.26.4
openai==1.30.3
packaging==24.0
pillow==10.3.0
pyautogen==0.2.27
pydantic==2.7.1
pydantic_core==2.18.2
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
pypdf==4.2.0
python-docx==1.1.2
python-dotenv==1.0.1
regex==2024.5.15
requests==2.32.2
sniffio==1.3.1
termcolor==2.4.0
tiktoken==0.7.0
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==2.2.1
